refactor(monitor): split monitor management into separate modules
- Move monitor information handling to new monitor_info module
- Create new wallpaper_manager module for wallpaper operations
- Update error handling to use new MonitorWallpaperManagerError type
- Refactor MonitorManager to use new modules
- Improve code organization and separation of concerns
